MA-1792 Ajout de test pour pas enlever un group d'un user qu'on a pas le droit

+++ Cause.SecurityManagement/Services/BaseUserManagementService.cs
@@ -11,28 +11,31 @@ namespace Cause.SecurityManagement.Services
 {
 	public class UserManagementService<TUser> : IUserManagementService<TUser> where TUser : User, new() 
 	{
-		private readonly IEmailForUserModificationSender emailSender;
+        private readonly IEmailForUserModificationSender emailSender;
         private readonly IUserGroupRepository userGroupRepository;
 		private readonly IUserPermissionRepository userPermissionRepository;
         private readonly IGroupPermissionRepository groupPermissionRepository;
         private readonly IUserRepository<TUser> userRepository;
+        private readonly IUserGroupPermissionService userGroupPermissionService;
 		protected readonly SecurityConfiguration SecurityConfiguration;
 
-		public UserManagementService(
-			IOptions<SecurityConfiguration> securityOptions,
+        public UserManagementService(
+            IOptions<SecurityConfiguration> securityOptions,
             IUserGroupRepository userGroupRepository,
             IUserPermissionRepository userPermissionRepository,
             IGroupPermissionRepository groupPermissionRepository,
 			IUserRepository<TUser> userRepository,
-			IEmailForUserModificationSender emailSender = null)
+            IUserGroupPermissionService userGroupPermissionService,
+            IEmailForUserModificationSender emailSender = null)
 		{
             this.emailSender = emailSender;
             this.userGroupRepository = userGroupRepository;
             this.userPermissionRepository = userPermissionRepository;
 			this.groupPermissionRepository = groupPermissionRepository;
             this.userRepository = userRepository;
+            this.userGroupPermissionService = userGroupPermissionService;
 
-			SecurityConfiguration = securityOptions.Value;
+            SecurityConfiguration = securityOptions.Value;
         }
 
 		public virtual List<TUser> GetActiveUsers()
@@ -90,7 +93,7 @@ namespace Cause.SecurityManagement.Services
             return userRepository.EmailIsAlreadyInUse(email, idUserToIgnore);
         }
 
-		public virtual void UpdateUserGroup(User user)
+        public virtual void UpdateUserGroup(User user)
         {
             if (user.Groups is null)
             {
@@ -102,7 +105,7 @@ namespace Cause.SecurityManagement.Services
 
             dbUserGroups.ForEach(userGroup =>
             {
-                if (userGroups.Any(g => g.Id == userGroup.Id) == false)
+                if (userGroups.Any(g => g.Id == userGroup.Id) == false && userGroupPermissionService.CurrentUserHasRequiredPermissionForGroupsAccess(userGroup.IdGroup))
                 {
                     userGroupRepository.Remove(userGroup);
                 }
